First draft of a nested, space-delimited list
This parser does almost nothing. It reads the tokens linearly, writes
them to a buffer, and prints them at the end of each expression.

There is not much new here (relative to the adder example) except the
handling of strings.

In the nest.l line

```
{num}   { yylval = strdup(yytext); return NUM;  }
```

note that I copy the yytext pointer, rather than `yylval = yytext`. This
required, since the yytext is a global pointer to a character array that
changes when each patter is evaluated.

The lexer line

```
.   { return yytext[0]; }
```

returns any character not recognized by a prior rule. I could have
defined the tokens EQ, LPAREN, and RPAREN for '=', '(', and ')',
respectively. But there is not a good reason to do so, since it just
creates extra code probably reduces readability.