Add symbol table and scope lookup functions
The symbol table is a simple list, it would be better to have a hash
table. But linear lookup times are acceptable for little languages, I
suppose.

I define a Scope struct that holds a link to its parent and a symbol
table. Looking up a variable involves looking first in the current
scope, then ascending to parent scope, and so forth, until a scope with
the variable name defined is found.

I also implemented a path look function, that allows access to named
scopes in the current scope.

Everything I wrote is borken. I have the basic framework, but it is full
of bugs and doesn't compile.