Fix bugs in 'scope' and write tokens
I fixed all compile-time bugs and resolved all grammar conflicts.
However, I have a runtime segfault. Not too surprising considering all
the untested wiring I have in the lookup functions. I'll fix this in the
next commit.

When debugging, it is extremely useful to be able to print the tokens.
But relying on adhoc printf statements is ugly. I've added a macro for
printing tokens and their values, while preserving whitespace.

For a piece of code like:

```
x = 0

scope A {
    x = 1
    y = 42
    scope B {
        x = 2
    }
}
```

It will automatically write the following to a `tok.log` file:

```
IDENTIFIER:x =  INT:0

SCOPE  VARIABLE:A  {
    IDENTIFIER:x =  INT:1
    IDENTIFIER:y =  INT:42
    SCOPE  VARIABLE:B  {
        IDENTIFIER:x =  INT:2
    }
}
```