Add table
The table will contain an entry for every expression in the mouse
script. Some will be named, such as manifolds within compositions and
compositions themselves. Others will be anonymous, such as effects. All
items in a table are typed and their values are stored in a union. A
table contains an entry {name, type, value} and a pointer to a container
with the next Entry (Table is a stack).

In mouse (and probably rat as well), only the COMPOSITION type is
recursive, being a Table itself.

I have written functions for adding to tables, joining tables, getting
elements from tables, recursively getting, and recursively getting via a
path.

I have also implemented Manifold and Effect structures.