Wired table into bison file (SEGFAULTS)
Running yyparse builds a table. This table, currently, is the sole
output of the parser. From it we can generate the RIL to build manifolds
and attach effects. However, it misses the connections between
manifolds, for this we will need either another data structure or a
major refactor existing ones ...