Fixed segfaults and polymorph
There are two basic kinds of sections in Rat. One that maps
right-hand-side things to the manifolds specified on the
left-hand-paths. The other is a common 1-to-1 assignment. I want to be
able to handle both types without undue code duplication, so I wrote a
method for each that takes type as an argument and feeds it into a
case-switch table.

Mouse only has one type in each category, but Rat has many. The new
infrastructure should scale well.