Print MIL in forward order
The tables are built backwards (FILO). But I need to execute the lines
in the order they were received, so I have to flip the list. This
requires a call to `table_reverse`, which recursively reverses
the given table (descending into compositions).

I don't like having to do all these reversals. The details of data
structure implementatino should not affect higher level code.

The Selection structures are also backwards, but I am not currently
reversing them. Reverse order is not incorrect, since all left hand
manifolds are being coupled to the same thing, however it needlessly
obfuscates the code.

Anyway, I ought to change datastructures to a FIFO.