Replaced tables with linked entry lists
Now Table is just `{Entry* head, Entry* tail}`. `Entry` now contains a
`Entry* next` pointer to the next entry. This standard linked list is
simpler and doesn't need reversals.

It may be worthwhile to replace Table with a general LL structure. I
could use this for Path and Selection as well.