Parse type string to nested functions
Parse the type `([(a,b,c)],(x->z))` into

```
tuple(array(tuple(a,b,c)),function(x,z))
```

This is actually pretty clean and easy. I prefer it to lrparsing (at the
moment). PEG has its perks. Not having to deal with ambiguity is one of
them. But I don't know the algorithm, so I hesitate to judge.