Clean up lrparsing example
Even after cleaning, the lrparsing is nowhere near as succinct as the
arpeggio implementation. Additionally, I have no idea how to cleanly
parse the produced tree. Of course, it is just a nested tuple, so I can
recurse through it and do pattern matching on the types, but this is a
far cry from the elegance of the arpeggio's Visitor pattern.