Add simple LIL writer
This is just a simple tree descent algorithm that links parents to
children in a directed graph. The result is a list of strings, which
described the link. It is general enough that the current function that
produces LIL could easily be replaced with one that produced, say DOT.

See the TODO's in Interpreter.hs, there is a lot lacking.