Cleaned up showTree
So what I was looking was present in the Data.Foldable module.

```
showTree :: (Tree -> Maybe String) -> Tree -> [String]
showTree f t =
     (DF.toList . f) t
  ++ (DF.concat . g) (return t >>= popChild)
  ++ (DF.concat . g) (return t >>= topChild)
  where
    g = CM.LiftM . showTree f
```

Damn there are a lot of functions in Haskell, reminds me of R, where
there is always a clever way to do it, obtainable if you have a god-like
vocabulary. But with Haskell it is easier -- static typing makes the
functions searchable (Hoogle).