Add error handling to `expr2tree`
However, I don't think the interpreter is the place to catch syntax
errors. It should be able to handle incorrect input (just so it can
match all cases), but it should not be in charge of syntax error message
generation, that is the job of the parser.

Some of the error messages I wrote will never even make it into the data
structure now, because they are silently ignored by the compiler ...